Use level 1 and 2 terminology for tags
The Navigation and presentation team has recently clarified terminology around
organising topic. We should start aligning it in the code.

We use ‘levels’ when talking about the location of a topic in a tree. This
language is also reflected in our Google Analytics tracking so that we can
understand which topic level is being used.
  Level 0 – all topics, for example all browse categories (currently 'root'
            in the code)
  Level 1 – topic, for example Driving and transport
  Level 2 – subtopic, for example Driving licences

‘Parent-child’ is often used interchangeably with topics and subtopics, where
the ‘parent’ is a topic and the ‘children’ are subtopics of that topic. It can
get confusing in hierarchies that go deeper than 2 levels. We should avoid
using ‘parent-child’ language in guidance and publishing apps.

Parent-child is still used to describe associations between individual tags,
because this language is used the DB.
Didn't change the scopes as they were only used for migrations.

This also removes #top_level_mainstream_browse_page? to avoid duplication, as