Allow Level 1 Specialist topics to be archived
It is currently only possible to archive level two specialist topics in
Collections Publisher. We need the UI to support archiving level one specialist
topics, as it would make the workflow for retiring the specialist topic tree
easier.

The link to archive will only be shown when Level 1 topic has no subtopics or
when all subtopics are archived.