Use content ids to present curated lists (aka groups)
When a content item is tagged to curated Mainstream browse or Specialist topic
(parent class Tag) and this content item’s slug changes the link disappears
from the topic page. This is because the curated lists/groups are based on
slugs. The “contents” slugs don’t get updated as a part of dependency
resolution in Publishing API.

We want to use content_ids instead as they are less volatile.

The “contents” key will be removed after the migration.