Don't email subscribers of retired topics if successor is a document collection
When we archive and redirect a specialist topic, we automatically unsubscribe all its existing
email subscribers and send them a notification.

As part of the work to retire specialist topics, we will be converting some into document collections.
The content of these pages will be so similar that we will migrate subscribers from the topic over to the new
document collection subscriber list. We do not need to tell users about this, as from their pov there is little change
to the emails they will be receiving.

This temp fix creates a new problem: when we are redirecting a specialist topic to an existing document collection the content
of the pages will be different enough that we will need to unsubscribe users and let them know.

Ideally we would know the topics that are being redirected to existing document collections, so that we could add an
exception to this code. But we don't currently have this information, so excluding all document collections is required for the time being.