Add Content Item model
This will replace the  object that's passed through to the tag archiver from the archival form, and will allow us to access additional information from the content item of the document that is chosen to redirect an archived tag to. This is part of work to retire specialist topics, and will no longer be needed once this work is complete.