Move some logic into the email alerts updater
To enable us to retire specialist topics, we want to add more flexibility in the email alert api logic when we redirect to another piece of content on GOV.UK - in some cases, rather than bulk unsubscribing, we want to bulk migrate email subscriptions. This is a preparatory commit, following the renaming of the email alerts unsubscriber to the email alerts updater.

The unsubscribe message body has been moved into this class to keep logic relating to unsubscribing together. Previously, this message was passed in as an optional parameter, but we were never actually using this optionality and instead always passed in the same body (with some variance based on the base path of the document we're redirecting to). Doing this required us to pass in the  object to the Email Updater, but since we will need to access this anway later for the logic that decides whether or not to unsubscribe, this makes sense.