Implement logic for bulk migrating email subscriptions
In order to enable work to retire specialist topics, we want to ensure that we can bulk migrate email subscriptions from specialist topics to document collections that have been created to replace them. In these cases, we can migrate without notifying the subscribers, since the content is similar enough to enable this via the rules of GDPR.

We can identify these cases as follows:

When a specialist topic is archived, and redirected to a document collection, if that document collections has a 'mapped_specialist_topic_content_id' field within its details hash that matches the content id of the specialist topic, this is a document collection that has specifically been created to replace the specialist topic. In all other cases, we should follow the normal bulk unsubscribing route, notifying the subscribers that they have been unsubscribed.