New criteria for bulk migration
This commit adds support for scenario 2 from the list below. Prior to this
commit, only scenarios 1 and 3 were handled.

1 If a specialist topic is being archived and redirected to a document
collection that was converted from a specialist topic it will have a
populated mapped_specialist_topic_content_id field. In this case we should migrate
email subscribers from the specialist topic, over to the document collection
list.

2 If a specialist topic is being archived and redirected to a converted document
collection which is one of a small number with a populated
taxonomy_topic_email_override field, we should migrate users to the subscriber
list for the taxonomy topic instead.

3 If neither mapped_specialist_topic_content_id nor taxonomy_topic_email_override field
are present, users should be bulk unsubscribed.