Update tag archiver
In 5fee75d86009d58a8241eee43fa7b64c5f87cd9a we updated the UI in this
application so that it's no longer possible to redirect specialist topics
to specialist topics, only to redirect to content_items. This is because
we are soon retiring the entire specialist topic tree.

The tag archiver specs were not updated to reflect this change, and still
tested the topic -> topic redirect behaviour.

This commit:
- Updates the specs
- Removes the unnecessary type check in the email alert updater class
- Injects the email_alerts_updater class to simplify the tests