Validate that a taxonomy topic is present in the content store
We should probably check that the basepath for the taxonomy topic is still a live
content item, before we ask email alert api to find_or_create a subscriber list for
the topic. This guards against a very unikely edge case where we redirect to a
document collection which retains a link to a topic that's been unpublished.