Add meta description for Coronavirus page
It is missing in the GOV.UK internal search results.

The content for the coronavirus landing page used to be pulled in
from two places, a [yaml file](https://github.com/alphagov/govuk-coronavirus-content/blob/master/content/coronavirus_landing_page.yml) in a github repo (now archived), and in collections publisher.

There is duplicate metadata hardcoded in collections, which should
eventually be removed.
https://github.com/alphagov/collections/blob/main/app/presenters/coronavirus_landing_page_presenter.rb#L32