Limit the connection_pool gem version to < 3
connection_pool is a transitive dependency pulled in by sidekiq, however
sidekiq v7 (which we are using) does not support connection_pool v3 but
does not pin it to < v3. Therefore, we need to pin it ourselves until we
update to sidekiq v8 (which does support connection_pool v3), or sidekiq
release an update to v7 that pins the connection_pool version.