Use get_live_content instead of filtering by version
We'd like to get rid of the version param from this API endpoint, and
it's cleaner in this case anyway to get the live version of a bit of
content rather than mucking with state history.

The state_history .select that I've deleted was pretty confusing in any
case - the "hash" parameter was actually a string, and only because in
ruby `"published"["published"]` returns `"published"` was the loop
working at all.